pyproject.toml
[project]
name = "robot-arm-chess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.11.0",
    "google-generativeai>=0.8.5",
    "openai>=1.75.0",
    "opencv-python>=4.11.0.86",
]
